fix namespace collisions
Still some object reference issues e.g. when click create new
