+++ GameLibrary/Models/Review.cs
@@ -18,11 +18,11 @@ namespace GameLibrary.Models
     {
         public int Id { get; set; }
         public int GameId { get; set; }
-        public int UserId { get; set; }
+        public Guid UserId { get; set; }
         public required string Content { get; set; }
         public int Rating { get; set; }
 
-        public Game Game { get; set; }
+        public Game? Game { get; set; }
         public required User User { get; set; }
     }
 }
