+++ GameLibrary/Pages/Admin/Delete.cshtml
@@ -1,12 +1,13 @@
 ﻿@page
-@model GameLibrary.Pages.Game.DetailsModel
+@model DeleteModel
 
 @{
-    ViewData["Title"] = "Details";
+    ViewData["Title"] = "Delete";
 }
 
-<h1>Details</h1>
+<h1>Delete</h1>
 
+<h3>Are you sure you want to delete this?</h3>
 <div>
     <h4>Game</h4>
     <hr />
@@ -43,8 +44,10 @@
             @Html.DisplayFor(model => model.Game.Description)
         </dd>
     </dl>
-</div>
-<div>
-    <a asp-page="./Edit" asp-route-id="@Model.Game.Id">Edit</a> |
-    <a asp-page="./Index">Back to List</a>
+    
+    <form method="post">
+        <input type="hidden" asp-for="Game.Id" />
+        <input type="submit" value="Delete" class="btn btn-danger" /> |
+        <a asp-page="./Index">Back to List</a>
+    </form>
 </div>
