+++ GameLibrary/Pages/Admin/Delete.cshtml.cs
@@ -1,0 +1,72 @@
+// Copyright 2024 Web.Tech. Group17
+//
+// Licensed under the Apache License, Version 2.0 (the "License"):
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using GameLibrary.Data;
+using GameLibrary.Models;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.EntityFrameworkCore;
+
+namespace GameLibrary.Pages.Admin;
+
+public class DeleteModel : PageModel
+{
+    private readonly ApplicationDbContext _context;
+
+    public DeleteModel(ApplicationDbContext context)
+    {
+        _context = context;
+    }
+
+    [BindProperty]
+    public Game? Game { get; set; }
+
+    public async Task<IActionResult> OnGetAsync(int? id)
+    {
+        if (id == null)
+        {
+            return NotFound();
+        }
+
+        var game = await _context.Games.FirstOrDefaultAsync(m => m.Id == id);
+
+        if (game == null)
+        {
+            return NotFound();
+        }
+        else
+        {
+            Game = game;
+        }
+        return Page();
+    }
+
+    public async Task<IActionResult> OnPostAsync(int? id)
+    {
+        if (id == null)
+        {
+            return NotFound();
+        }
+
+        var game = await _context.Games.FirstOrDefaultAsync(m => m.Id == id);
+        if (game != null)
+        {
+            Game = game;
+            _context.Games.Remove(Game);
+            await _context.SaveChangesAsync();
+        }
+
+        return RedirectToPage("./Index");
+    }
+}
