+++ GameLibrary/Pages/Admin/Details.cshtml
@@ -1,13 +1,12 @@
 ﻿@page
-@model GameLibrary.Pages.Game.DeleteModel
+@model DetailsModel
 
 @{
-    ViewData["Title"] = "Delete";
+    ViewData["Title"] = "Details";
 }
 
-<h1>Delete</h1>
+<h1>Details</h1>
 
-<h3>Are you sure you want to delete this?</h3>
 <div>
     <h4>Game</h4>
     <hr />
@@ -44,10 +43,8 @@
             @Html.DisplayFor(model => model.Game.Description)
         </dd>
     </dl>
-    
-    <form method="post">
-        <input type="hidden" asp-for="Game.Id" />
-        <input type="submit" value="Delete" class="btn btn-danger" /> |
-        <a asp-page="./Index">Back to List</a>
-    </form>
+</div>
+<div>
+    <a asp-page="./Edit" asp-route-id="@Model.Game.Id">Edit</a> |
+    <a asp-page="./Index">Back to List</a>
 </div>
