+++ GameLibrary/Pages/Admin/Edit.cshtml
@@ -1,11 +1,11 @@
 ﻿@page
-@model GameLibrary.Pages.Game.CreateModel
+@model EditModel
 
 @{
-    ViewData["Title"] = "Create";
+    ViewData["Title"] = "Edit";
 }
 
-<h1>Create</h1>
+<h1>Edit</h1>
 
 <h4>Game</h4>
 <hr />
@@ -13,6 +13,7 @@
     <div class="col-md-4">
         <form method="post">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
+            <input type="hidden" asp-for="Game.Id" />
             <div class="form-group">
                 <label asp-for="Game.Title" class="control-label"></label>
                 <input asp-for="Game.Title" class="form-control" />
@@ -40,14 +41,14 @@
                 <span asp-validation-for="Game.Description" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <input type="submit" value="Create" class="btn btn-primary" />
+                <input type="submit" value="Save" class="btn btn-primary" />
             </div>
         </form>
     </div>
 </div>
 
 <div>
-    <a asp-page="Index">Back to List</a>
+    <a asp-page="./Index">Back to List</a>
 </div>
 
 @section Scripts {
