+++ GameLibrary/Pages/Admin/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model GameLibrary.Pages.Game.IndexModel
+@model IndexModel
 
 @{
     ViewData["Title"] = "Index";
@@ -14,22 +14,22 @@
     <thead>
         <tr>
             <th>
-                @Html.DisplayNameFor(model => model.Game[0].Title)
+                @Html.DisplayNameFor(model => model.Games[0].Title)
             </th>
             <th>
-                @Html.DisplayNameFor(model => model.Game[0].Genre)
+                @Html.DisplayNameFor(model => model.Games[0].Genre)
             </th>
             <th>
-                @Html.DisplayNameFor(model => model.Game[0].ReleaseDate)
+                @Html.DisplayNameFor(model => model.Games[0].ReleaseDate)
             </th>
             <th>
-                @Html.DisplayNameFor(model => model.Game[0].Description)
+                @Html.DisplayNameFor(model => model.Games[0].Description)
             </th>
             <th></th>
         </tr>
     </thead>
     <tbody>
-@foreach (var item in Model.Game) {
+@foreach (var item in Model.Games) {
         <tr>
             <td>
                 @Html.DisplayFor(modelItem => item.Title)
