--- GameLibrary/Pages/Games/Create.cshtml.cs
@@ -1,45 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using GameLibrary.Data;
-using GameLibrary.Models;
-
-
-namespace GameLibrary.Pages.Game
-{
-    public class CreateModel : PageModel
-    {
-        private readonly GameLibrary.Data.ApplicationDbContext _context;
-
-        public CreateModel(GameLibrary.Data.ApplicationDbContext context)
-        {
-            _context = context;
-        }
-
-        public IActionResult OnGet()
-        {
-            return Page();
-        }
-
-        [BindProperty]
-        public GameLibrary.Models.Game Game { get; set; } = default!;
-
-        // For more information, see https://aka.ms/RazorPagesCRUD.
-        public async Task<IActionResult> OnPostAsync()
-        {
-            if (!ModelState.IsValid)
-            {
-                return Page();
-            }
-
-            _context.Game.Add(Game);
-            await _context.SaveChangesAsync();
-
-            return RedirectToPage("./Index");
-        }
-    }
-}
