--- GameLibrary/Pages/Games/Details.cshtml.cs
@@ -1,43 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
-using GameLibrary.Data;
-using GameLibrary.Models;
-
-namespace GameLibrary.Pages.Game
-{
-    public class DetailsModel : PageModel
-    {
-        private readonly GameLibrary.Data.ApplicationDbContext _context;
-
-        public DetailsModel(GameLibrary.Data.ApplicationDbContext context)
-        {
-            _context = context;
-        }
-
-        public GameLibrary.Models.Game Game { get; set; } = default!;
-
-        public async Task<IActionResult> OnGetAsync(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var game = await _context.Game.FirstOrDefaultAsync(m => m.Id == id);
-            if (game == null)
-            {
-                return NotFound();
-            }
-            else
-            {
-                Game = game;
-            }
-            return Page();
-        }
-    }
-}
